backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
python-decouple==3.8
requests==2.31.0
google-auth==2.25.2
